Bakalar-Mark-Vlkolensky-kody/Bakalar/MicroserviceEDA/RequestAPI/requirements.txt
aiofiles==0.8.0
aiohttp==3.7.4.post0
aioresponses==0.7.2
aiosqlite==0.17.0
amqp==5.1.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.4.0
billiard==3.6.4.0
blinker==1.4
celery==5.1.2
certifi==2021.10.8
chardet==4.0.0
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.0.3
Flask-WTF==0.15.1
greenlet==1.1.2
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==2.10
itsdangerous==2.1.1
Jinja2==3.0.3
kafka-python==2.0.2
kombu==5.2.4
MarkupSafe==2.1.1
multidict==6.0.2
priority==2.0.0
prompt-toolkit==3.0.28
pytz==2021.3
Quart==0.15.1
quart-auth==0.5.0
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.23
toml==0.10.2
typing-extensions==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
wsproto==1.1.0
WTForms==2.3.3
yarl==1.7.2
